+++ Domain/Interface/IProductRepository.cs
@@ -11,9 +11,9 @@ namespace Domain.Interface
     {
         Task<IList<Product>> GetallProducts();
         Task<Product> GetProduct(int productid);
-        Task ChangeProduct(int id, string productname, string imgurl, string shortdescription, string longdescription,int countitem,double price, string imgname);
+        Task ChangeProduct(int id, string productname, string imgurl, string shortdescription, string longdescription,int countitem,decimal price, string imgname);
         Task DeleteProduct(int productId);
-        Task AddProduct(string productname, string imgurl, string shortdescription, string longdescription,int countitem,double price,string imgname);
+        Task AddProduct(string productname, string imgurl, string shortdescription, string longdescription,int countitem,decimal price,string imgname);
         Task<IList<Product>> GetForIndex(int pageid);
         Task<int> CountOfProducs();
         Task AddViewCount(int productid);
